@@ -3,9 +3,8 @@
     
     https://github.com/jorisBarkema/svelte-burger-menu
 -->
-<script>
-    export let open;
-    export let duration;
+<script lang="ts">
+    let { open = $bindable(), duration } = $props();
 </script>
 
 <style>
@@ -44,7 +43,7 @@
     }
 </style>
 
-<button class:open on:click={() => open = !open}
+<button class:open onclick={() => open = !open}
     style="transition: color {duration}s ease-in-out; color: {open ? 'var(--color-bg)' : 'var(--color-primary)'};">
 	<svg width=32 height=32>
         <line id="top" x1=0 y1=9    x2=32 y2=9    style="transition: transform {duration}s ease-in-out, opacity {duration}s ease-in-out;"/>
