@@ -2,6 +2,10 @@
    <title>{title ? title + " - Herní archiv" : "Herní archiv"}</title>
 </svelte:head>
 
-<script>
-   export let title = ""
+<script lang="ts">
+   interface Props {
+      title?: string;
+   }
+
+   let { title = "" }: Props = $props();
 </script>
