@@ -1,12 +1,16 @@
 <script lang="ts">
 	import type { Person } from '$src/types';
 
-    export let person: Person;
+    interface Props {
+        person: Person;
+    }
+
+    let { person }: Props = $props();
 
 
 </script>
 
 { person.name }
-{#if person.shortname }
+{#if person.shortname}
     (<span style="color: {person.color || 'default'};">{ person.shortname }</span>)
 {/if}
