@@ -1,3 +1,4 @@
+<!-- @migration-task Error while migrating Svelte code: Unexpected block closing tag -->
 <script lang="ts">
 	import type { InterviewData } from '$src/types';
     import { tocCrawler } from './TableOfContents/crawler';
@@ -9,108 +10,113 @@
 </script>
 
 <Post title={data.title}>
-    <div slot="side">
-        {#if data.narrator.photo }
-            <img src="{ data.narrator.photo.url }" alt="Fotografie narátora" />
-        {/if}
-        <div>
-            <strong>
-                { data.narrator.name }
-            </strong>
-        </div>
-        {#if data.interview.project }
-            { data.interview.project.name }
-        {/if}
-
-        <hr>
-
-        <TableOfContents />
-    </div>
-    <div slot="content" class="content">
-        <h2>{ data.title }</h2>
-        <InterviewAudio {data} />
-        <dl>
-            <dt>Jméno { data.narrator.gender == "M" ? "narátora" : "narátorky" }</dt>
-            <dd><NameWithShortname person={data.narrator} /></dd>
-
-            <dt>Ročník narození</dt>
-            <dd>{ data.narrator.birth_year }</dd>
-
-            <dt>Datum rozhovoru</dt>
-            <dd>{ data.interview.date.toLocaleDateString("cs-CZ") }</dd>
-
-            <dt>Místo rozhovoru</dt>
-            <dd>{ data.interview.place }</dd>
-            
-            {#if data.interview.interviewers }
-                <dt>Jméno tazatelů</dt>
-                <dd>
-                    {#each data.interview.interviewers as person, i}
-                        {#if i > 0},<br>{/if}
-                        <NameWithShortname {person} />
-                    {/each}
-                </dd>
+    {#snippet side()}
+        <div class="side">
+            {#if data.narrator.photo }
+                <img src="{ data.narrator.photo.url }" alt="Fotografie narátora" />
             {/if}
-
-            {#if data.interview.interviewer }
-                <dt>Jméno tazatele</dt>
-                <dd><NameWithShortname person={data.interview.interviewer} /></dd>
-            {/if}
-
-            <dt>Délka rozhovoru</dt>
-            <dd>{ data.interview.length }</dd>
-
-            
+            <div>
+                <strong>
+                    { data.narrator.name }
+                </strong>
+            </div>
             {#if data.interview.project }
-                <dt>Projekt</dt>
-                <dd>
-                    {#if data.interview.project.url}
-                        <a href={data.interview.project.url}>{ data.interview.project.name }</a>
-                    {:else}
-                        { data.interview.project.name }
-                    {/if}
-                </dd>
+                { data.interview.project.name }
             {/if}
 
-            {#if data.interview.transcriber }
-                <dt>Přepis</dt>
-                <dd>{ data.interview.transcriber.name }</dd>
+            <hr>
+
+            <TableOfContents />
+        </div>
+    {/snippet}
+    {#snippet content()}
+        <article>
+            <h2>{ data.title }</h2>
+            <InterviewAudio {data} />
+            <dl>
+                <dt>Jméno { data.narrator.gender == "M" ? "narátora" : "narátorky" }</dt>
+                <dd><NameWithShortname person={data.narrator} /></dd>
+
+                <dt>Ročník narození</dt>
+                <dd>{ data.narrator.birth_year }</dd>
+
+                <dt>Datum rozhovoru</dt>
+                <dd>{ data.interview.date.toLocaleDateString("cs-CZ") }</dd>
+
+                <dt>Místo rozhovoru</dt>
+                <dd>{ data.interview.place }</dd>
+                
+                {#if data.interview.interviewers }
+                    <dt>Jméno tazatelů</dt>
+                    <dd>
+                        {#each data.interview.interviewers as person, i}
+                            {#if i > 0},<br>{/if}
+                            <NameWithShortname {person} />
+                        {/each}
+                    </dd>
+                {/if}
+
+                {#if data.interview.interviewer }
+                    <dt>Jméno tazatele</dt>
+                    <dd><NameWithShortname person={data.interview.interviewer} /></dd>
+                {/if}
+
+                <dt>Délka rozhovoru</dt>
+                <dd>{ data.interview.length }</dd>
+
+                
+                {#if data.interview.project }
+                    <dt>Projekt</dt>
+                    <dd>
+                        {#if data.interview.project.url}
+                            <a href={data.interview.project.url}>{ data.interview.project.name }</a>
+                        {:else}
+                            { data.interview.project.name }
+                        {/if}
+                    </dd>
+                {/if}
+
+                {#if data.interview.transcriber }
+                    <dt>Přepis</dt>
+                    <dd>{ data.interview.transcriber.name }</dd>
+                {/if}
+
+                {#if data.interview.redaction }
+                    <dt>Redakce</dt>
+                    <dd>{ data.interview.redaction.name }</dd>
+                {/if}
+
+                {#if data.interview.publication_date }
+                    <dt>Datum uveřejnění</dt>
+                    <dd>{ data.interview.publication_date.toLocaleDateString("cs-CZ") }</dd>
+                {/if}
+            </dl>
+
+            {#if data.narrator.bio }
+                <p>{@html data.narrator.bio }</p>
             {/if}
 
             {#if data.interview.redaction }
-                <dt>Redakce</dt>
-                <dd>{ data.interview.redaction.name }</dd>
+                <p>Poznámka: Text níže je upravená verze kompletního přepisu rozhovoru. Pokud máte z výzkumných nebo jiných důvodů zájem o kompletní přepis, napište na <a href="mailto:info@hernihistorie.cz">info@hernihistorie.cz</a> nebo <a href="mailto:sucharmail@gmail.com">sucharmail@gmail.com</a>.
+                </p>
             {/if}
 
-            {#if data.interview.publication_date }
-                <dt>Datum uveřejnění</dt>
-                <dd>{ data.interview.publication_date.toLocaleDateString("cs-CZ") }</dd>
+            <hr>
+            
+            <article id="tocTarget" use:tocCrawler={{ mode: 'generate' }}>
+                <slot name="content"/>
+            </article>
+
+            <hr>
+
+            {#if data.narrator.photo && data.narrator.photo.license_text }
+                <h4>Fotografie</h4>
+                <a href="{ data.narrator.photo.details_url }">
+                    { data.narrator.photo.license_text }
+                </a>
             {/if}
-        </dl>
-
-        {#if data.narrator.bio }
-            <p>{@html data.narrator.bio }</p>
-        {/if}
-
-        {#if data.interview.redaction }
-            <p>Poznámka: Text níže je upravená verze kompletního přepisu rozhovoru. Pokud máte z výzkumných nebo jiných důvodů zájem o kompletní přepis, napište na <a href="mailto:info@hernihistorie.cz">info@hernihistorie.cz</a> nebo <a href="mailto:sucharmail@gmail.com">sucharmail@gmail.com</a>.
-        {/if}
-
-        <hr>
-        
-        <article id="tocTarget" use:tocCrawler={{ mode: 'generate' }}>
-            <slot name="content"/>
         </article>
-
-        <hr>
-
-        {#if data.narrator.photo && data.narrator.photo.license_text }
-            <h4>Fotografie</h4>
-            <a href="{ data.narrator.photo.details_url }">
-                { data.narrator.photo.license_text }
-            </a>
-        {/if}
-    </div>
+    {/snippet}
 </Post>
 
 <style>
@@ -118,7 +124,7 @@
         margin-bottom: 0;
     }
 
-    div[slot="side"] {
+    div.side {
         position: sticky;
         top: 2em;
 
@@ -128,7 +134,7 @@
         overflow-y: scroll;
     }
 
-    div[slot="side"] img {
+    div.side img {
         border-radius: 8px;
     }
 
@@ -141,7 +147,7 @@
         }
     }
 
-    .content :global(hr) {
+    article :global(hr) {
         margin-top: 3em;
         margin-bottom: 3em;
     }
